https://gitlab.com/SQAD-MT/Web/WEBPro/issues/1103
* fix percents parsing
* fix boundaries of totals calculating - include lower bound

+++ src/WebApiContrib.Formatting.Xlsx/Serialisation/Views/Formatted/FormattedExcelDataRow.cs
@@ -38,7 +38,7 @@ namespace SQAD.MTNext.WebApiContrib.Formatting.Xlsx.Serialisation.Views.Formatte
             {
                 if (isPercent)
                 {
-                    return longResult / 100;
+                    return (decimal) longResult / 100;
                 }
 
                 return longResult;
